+++ src/store/modules/snapshots.js
@@ -1,10 +1,9 @@
-import { DEFAULT_DATA_TABLE_OPTIONS } from '../../consts'
+import { DEFAULT_DATA_TABLE_OPTIONS } from '@/consts'
 
 export const snapshots = {
   namespaced: true,
   state: {
     filter: '',
-    repository: '',
     pagination: Object.assign({}, DEFAULT_DATA_TABLE_OPTIONS, { sortBy: ['name'] })
   },
   mutations: {
@@ -13,9 +12,6 @@ export const snapshots = {
     },
     setPagination (state, pagination) {
       state.pagination = pagination
-    },
-    setRepository (state, repository) {
-      state.repository = repository
     }
   }
 }