+++ src/views/Snapshots.vue
@@ -1,43 +1,51 @@
 <template>
   <v-card>
     <v-card-title>
-      <h2 class="text-h5">{{ $t('snapshots.heading') }}</h2>
-      <reload-button :action="reloadData"/>
+      <h2 class="text-h5">{{ $t('snapshots.heading', { name: 'dirk' }) }}</h2>
+      <reload-button :action="load"/>
     </v-card-title>
     <v-divider/>
 
-    <snapshots-table-wrapper ref="tableWrapper"/>
-
-    <v-card-text>
-      <div class="d-inline-flex">
-        <path-repo-info/>
-      </div>
-    </v-card-text>
+    <loader :request-state="requestState" hide-progress>
+      <snapshots-table :snapshots="data || []"
+                       :loading="requestState.loading"
+                       :repository="repository"
+                       @reloadData="load"/>
+    </loader>
   </v-card>
 </template>
 
 <script>
   import ReloadButton from '@/components/shared/ReloadButton'
-  import SnapshotsTableWrapper from '@/components/Snapshots/SnapshotsTableWrapper'
-  import PathRepoInfo from '@/components/shared/PathRepoInfo'
-  import { ref } from 'vue'
+  import Loader from '@/components/shared/Loader'
+  import SnapshotsTable from '@/components/Snapshots/SnapshotsTable'
+  import { setupElasticsearchRequest } from '@/mixins/RequestComposition'
+  import { useRouter } from '@/helpers/composition'
+  import { onMounted } from 'vue'
 
   export default {
     name: 'snapshots',
     components: {
       ReloadButton,
-      SnapshotsTableWrapper,
-      PathRepoInfo
+      SnapshotsTable,
+      Loader
     },
     setup () {
-      const tableWrapper = ref(null)
-      const reloadData = () => {
-        tableWrapper.value.loadSnapshots()
-      }
+      const { route } = useRouter()
+
+      const repository = route.params.repositoryName
+      const {
+        load,
+        requestState,
+        data
+      } = setupElasticsearchRequest('catSnapshots', { repository })
+      onMounted(load)
 
       return {
-        tableWrapper,
-        reloadData
+        requestState,
+        load,
+        data,
+        repository
       }
     }
   }