+++ tests/e2e/specs/views/snapshots.js
@@ -64,7 +64,7 @@ describe('Snapshots', () => {
     beforeEach(() => {
       cy.createSnapshotRepository(repositoryName)
       cy.get('#reload-snapshot-repositories').click()
-      cy.get('tr').contains(repositoryName).click()
+      cy.get('a').contains('Manage Snapshots').click()
     })
 
     it('can create a new snapshot with default settings', () => {
@@ -73,15 +73,5 @@ describe('Snapshots', () => {
       cy.get('#snapshot_name').type(snapshotName)
       cy.get('#create_snapshot').click()
     })
-
-    // it('can delete a snapshot', () => {
-    //   let snapshotName = 'snap-1'
-    //   cy.createNewSnapshot(repositoryName, snapshotName)
-    //   cy.get('#reload-button').click()
-    //   cy.get('tr').contains(snapshotName).get('button[title="Options"]').click()
-    //   cy.get('tr').contains(snapshotName).get('div.v-list-item__content').contains('Delete snapshot').click()
-    //   cy.get('#reload-button').click()
-    //   cy.get('tr').contains(snapshotName).should('not.exist')
-    // })
   })
 })